Library (error): add the gal_error function
Until now, developers had to write quite a few lines to create an error
message. This included creating the message while checking for any possible
errors, and finally adding it to the error queue. See the code below for an
example of the previous implementation:

  void example(int arg, gal_error_t **err)
  {
    char *errstr=NULL;
    if ( arg < 0 )
      if(asprintf(&errstr,
                  "%s: arg must be positive '%d'", __func__, arg) < 0)
        gal_checkset_malloc_cat((char *)__func__, ": can't use `asprintf`");
      gal_error_add_back_msg(err, errstr, GAL_EXAMPLE_ERROR);
  }

With this commit, we've added the new 'gal_error' function which cleans up
the source code as shown below:

void example(int arg, gal_error_t **err)
  {
    if ( arg < 0 )
      gal_error(err, GAL_EXAMPLE_ERROR,
                "%s: arg must be positive '%d'", __func__, arg);
  }

As can be seen, the 'errstr', 'asprintf', 'gal_error_add_back_msg',
'gal_checkset_malloc_cat' and the conditional are not needed now. This
allows for easier and less error-prone implementation of error messages.