Library (errorinprogram.h): included for usage in programs
Until now, we included 'gnuastro-internal/error-internal' in the programs
that should print and abort with an error from the library. However, this
file was not actually committed in the version-controlled history!

With this commit, a new version of this file has been put in
'lib/gnuastro/errorinprogram.h' since this file should actually be
installed with Gnuastro and accessible to users of the Gnuastro library!
But it isn't a normal header file because it contains the full definition
of the function (because it should not be compiled into the library due to
the 'exit' command).

Besides these, I also made the following corrections:

 - Fixed line widths in part that they exceeded the standard 75 characters.

 - The error messages now include the library name rather than the library
   code (which is hard to interpret for any reader; even a developer). For
   this, the 'gal_error_print_lib_name' has been defined and a simple shell
   command has been added in the comments of the 'gal_error_library_codes'
   enum.

 - Within the 'lib/error.c' functions, all checks have been brought after
   the variable definitions (which is standard Gnuastro practice).

IN PROGRESS: see why the function names are not being printed in the
example commadn of 'tests/during-dev.sh'.