Library (error.h): option for enabling verbose errors
Until now, in this development branch, we were only printing a very verbose
and detailed error message that may be too much information for a
normal/new user (thus causing more confusion!).

With this commit, a new common option has been added for all the programs
called '--verbose-errors' to manage this. The erorr printing function has
also been split into two functions: one to only print the string, and one
to print the full error message to standard output. The verbose argument
has been added to the arguments of these functions, but it still doesn't do
anything.

This was done after a discussion with Labeeb Asari and Pedram
Ashofteh-Ardakani.