Added Python Error Types
Until Now, the error mechnaism only retained about the library and function which caused the error. However while reporting errors in python, we also have to specify the Type of error that has occured.
With this commit, A new enum structure has been added which lists down possible error types in python. The programmer can classify each error into one of the error types of python so the python interface can invoke the respective error.